client

+++ resources/client/clients.py
@@ -16,18 +16,23 @@ class ClsClient(Resource):
     def get(self):
         try:
             conn = mysql.connector.connect(**config)
+            cursor = conn.cursor(dictionary=True)
             client_id = int(request.headers.get('client_id'))
-            if not isinstance(conn, str):
-                cursor = conn.cursor(dictionary=True)
-                query = "SELECT client_id, client_name, client_status, mobile_phone, email_address, city FROM client \
-                         WHERE client_id = {client_id}".format(client_id = client_id)
-                cursor.execute(query)
-                clientRow = cursor.fetchone()
-                close_connection(conn, cursor)
+            cursor.execute("SELECT count(*) FROM client WHERE client_id=%s", (client_id,))
+            check_record_exist = cursor.fetchone().get('count(*)')
+            if check_record_exist:
+                if not isinstance(conn, str):
+                    query = "SELECT client_id, client_name, client_status, mobile_phone, email_address, city FROM client \
+                             WHERE client_id = {client_id}".format(client_id = client_id)
+                    cursor.execute(query)
+                    clientRow = cursor.fetchone()
+                    close_connection(conn, cursor)
 
-                return api_success(clientRow, "Client details fetched successfully")
+                    return api_success(clientRow, "Client details fetched successfully")
+                else:
+                    return api_failure(str(conn))
             else:
-                return api_failure(str(conn))
+                return { "message": "Record not found" }, 404
         except Exception as error:
             return api_failure(str(error))
 
@@ -60,6 +65,7 @@ class ClsClient(Resource):
     def put(self):
         try:
             conn = mysql.connector.connect(**config)
+            cursor = conn.cursor(dictionary=True)
             _json = request.json
             _client_id = _json['client_id']
             _client_name = _json['client_name']
@@ -67,18 +73,23 @@ class ClsClient(Resource):
             _mobile_phone = _json['mobile_phone']
             _email_address = _json['email_address']
             _city = _json['city']
-            if not isinstance(conn, str):
-                cursor = conn.cursor(dictionary=True)
-                query = "UPDATE client set client_name=%s, client_status=%s, mobile_phone=%s, email_address=%s, city=%s where client_id = %s"
-                bindData = (_client_name, _client_status, _mobile_phone, _email_address, _city, _client_id)
-                cursor = conn.cursor()
-                cursor.execute(query, bindData)
-                conn.commit()
-                close_connection(conn, cursor)
 
-                return api_success(None, "Client updated successfully")
+            cursor.execute("SELECT count(*) FROM client WHERE client_id=%s", (_client_id,))
+            check_record_exist = cursor.fetchone().get('count(*)')
+            if check_record_exist:
+                if not isinstance(conn, str):
+                    query = "UPDATE client set client_name=%s, client_status=%s, mobile_phone=%s, email_address=%s, city=%s where client_id = %s"
+                    bindData = (_client_name, _client_status, _mobile_phone, _email_address, _city, _client_id)
+                    cursor = conn.cursor()
+                    cursor.execute(query, bindData)
+                    conn.commit()
+                    close_connection(conn, cursor)
+
+                    return api_success(None, "Client updated successfully")
+                else:
+                    return api_failure(str(conn))
             else:
-                return api_failure(str(conn))
+                return { "message": "Record not found" }, 404
         except Exception as error:
             return api_failure(str(error))
 
@@ -86,17 +97,22 @@ class ClsClient(Resource):
     def delete(self):
         try:
             conn = mysql.connector.connect(**config)
+            cursor = conn.cursor(dictionary=True)
             _client_id = int(request.headers.get('client_id'))
-            if not isinstance(conn, str):
-                cursor = conn.cursor(dictionary=True)
-                query = "DELETE FROM client WHERE client_id =%s"
-                bindData = (_client_id, )
-                cursor.execute(query, bindData)
-                conn.commit()
-                close_connection(conn, cursor)
+            cursor.execute("SELECT count(*) FROM client WHERE client_id=%s", (_client_id,))
+            check_record_exist = cursor.fetchone().get('count(*)')
+            if check_record_exist:
+                if not isinstance(conn, str):
+                    query = "DELETE FROM client WHERE client_id =%s"
+                    bindData = (_client_id, )
+                    cursor.execute(query, bindData)
+                    conn.commit()
+                    close_connection(conn, cursor)
 
-                return api_success(None, "Client deleted successfully")
+                    return api_success(None, "Client deleted successfully")
+                else:
+                    return api_failure(str(conn))
             else:
-                return api_failure(str(conn))
+                return { "message": "Record not found" }, 404
         except Exception as error:
             return api_failure(str(error))
